separate yolov8 model file and tests
